Remove transaction described entity
