+++ SUMMARY.md
@@ -1,22 +1,21 @@
 # Table of contents
 
-* [⚡ Aragon Connect](README.md)
+- [⚡ Aragon Connect](README.md)
 
 ## GUIDES
 
-* [Getting started](guides/getting-started.md)
-* [Aragon app subgraph](guides/subgraph-template.md)
+- [Getting started](guides/getting-started.md)
+- [Aragon app subgraph](guides/subgraph-template.md)
 
 ## API REFERENCE
 
-* [connect\(\)](api-reference/connect.md)
-* [App](api-reference/app.md)
-* [Connectors](api-reference/connectors.md)
-* [Intent](api-reference/transaction-intent.md)
-* [Organization](api-reference/organization.md)
-* [Permission](api-reference/permission.md)
-* [Repo](api-reference/repo.md)
-* [Role](api-reference/role.md)
-* [TransactionPath](api-reference/transaction-path.md)
-* [TransactionRequest](api-reference/transaction-request.md)
-
+- [connect\(\)](api-reference/connect.md)
+- [App](api-reference/app.md)
+- [Connectors](api-reference/connectors.md)
+- [Intent](api-reference/transaction-intent.md)
+- [Organization](api-reference/organization.md)
+- [Permission](api-reference/permission.md)
+- [Repo](api-reference/repo.md)
+- [Role](api-reference/role.md)
+- [TransactionPath](api-reference/transaction-path.md)
+- [TransactionRequest](api-reference/transaction-request.md)
