+++ api-reference/transaction-request.md
@@ -4,8 +4,10 @@ An object describing a transaction that can get signed by a library like ethers.
 
 ## Properties
 
-| Name   | Type     | Description                       |
-| :----- | :------- | :-------------------------------- |
-| `data` | `String` | Transaction data.                 |
-| `from` | `String` | Address to use as default sender. |
-| `to`   | `String` | Target address or ENS name.       |
+| Name                   | Type           | Description                                                                                  |
+| :--------------------- | :------------- | :------------------------------------------------------------------------------------------- |
+| `annotatedDescription` | `Annotation[]` | List of the Radspec description bindings with the properties `{ type: string, value: any }`. |
+| `description`          | `string`       | Radspec description for the transaction.                                                     |
+| `data`                 | `String`       | Transaction data.                                                                            |
+| `from`                 | `String`       | Address to use as default sender.                                                            |
+| `to`                   | `String`       | Target address or ENS name.                                                                  |
