+++ packages/connect-core/src/utils/radspec/index.ts
@@ -5,7 +5,6 @@ import { addressesEqual } from '../address'
 import { findAppMethodFromIntent } from '../app'
 import Application from '../../entities/Application'
 import { TransactionRequestData } from '../../transactions/TransactionRequest'
-import { TransactionDescribedData } from '../../transactions/TransactionDescribed'
 import { Abi, AppIntent } from '../../types'
 
 interface FoundMethod {
@@ -20,7 +19,7 @@ export async function tryEvaluatingRadspec(
   intent: TransactionRequestData,
   apps: Application[],
   provider?: ethers.providers.Provider // Decorated intent with description, if one could be made
-): Promise<TransactionDescribedData> {
+): Promise<TransactionRequestData> {
   const app = apps.find(app => addressesEqual(app.address, intent.to))
 
   // If the intent matches an installed app, use only that app to search for a
